refactor(agents): extract attempt message summary helpers (#72072)
